chore: reorganize scripts into smoke and runs(e2e)
