+++ scripts/runs/run_eval_artifacts.py
@@ -1,5 +1,5 @@
 """CLI wrapper for generating eval artifacts from evidence previews."""
-# Run: python scripts/run_eval_artifacts.py --config config/run_config.yaml --mode fixture_only
+# Run: python scripts/runs/run_eval_artifacts.py --config config/run_config.yaml --mode fixture_only
 
 from __future__ import annotations
 
@@ -11,7 +11,7 @@ if __package__ is None or __package__ == "":
     if str(PROJECT_ROOT) not in sys.path:
         sys.path.insert(0, str(PROJECT_ROOT))
 
-from scripts.run_evidence_preview import app
+from scripts.runs.run_evidence_preview import app
 
 
 if __name__ == "__main__":
