+++ scripts/smoke/baseline_eval.py
@@ -15,7 +15,7 @@ from time import perf_counter
 from typing import Any
 
 if __package__ is None or __package__ == "":
-    # Allow running as `python scripts/baseline_eval.py` by adding repo root.
+    # Allow running as `python scripts/smoke/baseline_eval.py` by adding repo root.
     PROJECT_ROOT = Path(__file__).resolve().parent.parent
     if str(PROJECT_ROOT) not in sys.path:
         sys.path.insert(0, str(PROJECT_ROOT))
