+++ scripts/smoke/planner_smoke_test.py
